Confirmation code validation and text update.
Changed confirmation code validation to be case insensitive by converting the code to lower case (codes in the database are stored in lower case).  Changed the text of the confirmation code printout to explicitly show the point value and how to redeem the code.  Also added text to the code describing the social bonus if the event/activity has a social bonus.