Removed avatar cache and changed outside of competition template.
The avatar cache is not smart enough to invalidate the thumbnails of the user's avatar, so the cache is temporarily removed.  This should fix #444, but it needs verification.  The text of the home/restricted.html template was changed to notify the user that their login was successful, but the competition hasn't started yet.  If the user tries to go to this page when the competition starts, they are redirected to the home page.