Fixed admin form issue and added styles to popular_tasks.
The admin form appears to store the last used fields, so after viewing an activity with variable points, every activity member has that field on their form and cannot be approved.  The fields are now set if the activity does not have variable points.

Popular tasks does not take up the entire table, so I added some styling to the first column to increase the width.