Updated 404 and 500 template.
Updated the 404 template to take them to the root index (and be redirected appropriately) instead of landing page.  Updated the 500 template with the raw contents of the 404 page.  This fixes a long standing issue where 500 errors would return something like 'KeyError: STATIC_URL does not exist'.  It turns out that reporting a 500 may not go through the context processors that insert these template variables.