Canopy karma not awarded to group members.
Implemented test cases for when a group member has accepted the mission already and when an activity is submitted before a user accepts the mission.  Changed canopy group email handling to only create activity members if the user is participating in its related mission.  MissionMember also checks if there are outstanding ActivityMembers that need to be added to the user.  Close #484.